refactor: Cvc 커스텀훅으로 도메인로직 분리
